Migrate CoreTests to addSourceLines
Replace the 9 remaining deprecated addSourceFile calls in CoreTests with
addSourceLines, inlining the test sources and stripping their license
headers. Removes the 8 now-unneeded @SuppressWarnings("deprecation")
annotations and deletes the 7 testdata files that no longer have users.

Util.java is shared by two tests, so its contents are factored into a
single UTIL_SOURCE constant. The file itself is kept, since other test
classes resolve it by fully-qualified name from the classpath.

Partial progress on #1425.
