--- nullaway/src/test/resources/com/uber/nullaway/testdata/CheckAssertSupportPositiveCases.java
@@ -1,46 +1,0 @@
-/*
- * Copyright (c) 2017 Uber Technologies, Inc.
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- * THE SOFTWARE.
- */
-
-package com.uber.nullaway.testdata;
-
-import javax.annotation.Nullable;
-
-public class CheckAssertSupportPositiveCases {
-
-  class T1 {
-    @Nullable Object obj;
-  }
-
-  void someMethod() {
-    T1 t1 = new T1();
-    assert t1.obj == null;
-    // BUG: Diagnostic contains: dereferenced expression
-    t1.obj.toString();
-  }
-
-  void someMethod2() {
-    T1 t1 = new T1();
-    assert t1 != null;
-    // BUG: Diagnostic contains: dereferenced expression
-    t1.obj.toString();
-  }
-}
